apps/fastapi-rapidjson/requirements.txt
fastapi==0.68.1
python-rapidjson==1.4
python-multipart==0.0.5
